Add new fields and adjust to backend output
An institution is made of the name, the slug (which identifies it) and
the thumbnail link.
The current_language field shows in what language the current
accommodation object is displayed (The translation falls back to the
default language if there is no entry in the current language saved in
the cms).
spoken_language is a list of objects containing the translated name, the
language code and native writing of the language.